backend/requirements.txt
fastapi
uvicorn
websockets
pydantic
python-multipart
google-generativeai
python-dotenv
